+++ src/Client/OptionsFactory.php
@@ -30,7 +30,6 @@ class OptionsFactory implements FactoryInterface
 
         $result = new Options();
         $result->setApiUrl($optionsConfig[ConfigKey::OPTION_API_URL] ?? '')
-               ->setAgent($optionsConfig[ConfigKey::OPTION_AGENT] ?? '')
                ->setAccessKey($optionsConfig[ConfigKey::OPTION_ACCESS_KEY] ?? '')
                ->setTimeout($optionsConfig[ConfigKey::OPTION_TIMEOUT] ?? 0);