+++ test/serializer/Entity/RecipeWithExpensiveVersionTest.php
@@ -82,14 +82,14 @@ class RecipeWithExpensiveVersionTest extends SerializerTestCase
     {
         return [
             'name' => 'ijk',
-            'label'=> 'lmn',
+            'label' => 'lmn',
             'description' => 'opq',
             'mode' => 'rst',
             'ingredients' => [
                 [
                     'type' => 'mno',
                     'name' => 'pqr',
-                    'label'=> 'stu',
+                    'label' => 'stu',
                     'description' => 'vwx',
                     'amount' => 2.3,
                 ],
@@ -98,7 +98,7 @@ class RecipeWithExpensiveVersionTest extends SerializerTestCase
                 [
                     'type' => 'klm',
                     'name' => 'nop',
-                    'label'=> 'qrs',
+                    'label' => 'qrs',
                     'description' => 'tuv',
                     'amount' => 4.5,
                 ],
@@ -106,14 +106,14 @@ class RecipeWithExpensiveVersionTest extends SerializerTestCase
             'craftingTime' => 73.31,
             'expensiveVersion' => [
                 'name' => 'wxy',
-                'label'=> 'zab',
+                'label' => 'zab',
                 'description' => 'cde',
                 'mode' => 'fgh',
                 'ingredients' => [
                     [
                         'type' => 'abc',
                         'name' => 'def',
-                        'label'=> 'ghi',
+                        'label' => 'ghi',
                         'description' => 'jkl',
                         'amount' => 1.2,
                     ],
@@ -122,7 +122,7 @@ class RecipeWithExpensiveVersionTest extends SerializerTestCase
                     [
                         'type' => 'yza',
                         'name' => 'bcd',
-                        'label'=> 'efg',
+                        'label' => 'efg',
                         'description' => 'hij',
                         'amount' => 3.4,
                     ],