+++ test/src/ApiClientFactoryTest.php
@@ -15,7 +15,6 @@ use Interop\Container\ContainerInterface;
 use JMS\Serializer\SerializerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use ReflectionException;
 
 /**
  * The PHPUnit test of the ApiClientFactory class.
@@ -30,7 +29,6 @@ class ApiClientFactoryTest extends TestCase
 
     /**
      * Tests the invoking.
-     * @throws ReflectionException
      * @covers ::__invoke
      */
     public function testInvoke(): void